commands/requirements.txt
discord.py
aiohttp
easy-pil